moving function to new file
