+++ aisquared/serving/__init__.py
@@ -4,4 +4,4 @@ from mlflow.pytorch import save_model as save_pytorch_model
 from mlflow.keras import save_model as save_keras_model
 
 from .deploy import deploy_model, load_mann_model, load_keras_model, load_pytorch_model, load_sklearn_model, load_tensorflow_model
-from .predict import get_remote_prediction
+from .get_remote_prediction import get_remote_prediction
