+++ aisquared/serving/get_remote_prediction.py
